requirements.txt
black
build
click
elasticsearch>=8.2.0
eql>=0.9.12
isort
nbformat
pytest
pytoml
requests
ruamel.yaml
semver
virtualenv
